Official BC CDC Fish Status Official BC Conservation Data Centre (CDC) fish data.

The search criteria were Fish, Freshwater OR Fish, Marine and the sort order was Scientific Name Ascending.

cdc

Format

A tbl data frame that includes:

Species Code

A character vector of the unique data centre fish code.

BC List

A character vector of the BC Conservation Listing.

Details

The data is released under the Open Government Licence - British Columbia

References

B.C. Conservation Data Centre. 2019. BC Species and Ecosystems Explorer. B.C. Minist. of Environ. Victoria, B.C. Available: http://a100.gov.bc.ca/pub/eswp/ (accessed Aug 23, 2019).

See also

Examples

head(cdc)

#> # A tibble: 6 x 45
#>   `Scientific Nam… `Scientific Nam… `English Name` `English Name S…
#>   <chr>            <chr>            <chr>          <chr>           
#> 1 Acantholiparis … NA               Fanged Snailf… NA              
#> 2 Acantholiparis … NA               Spiny Snailfi… NA              
#> 3 Acipenser medir… NA               Green Sturgeon NA              
#> 4 Acipenser trans… NA               White Sturgeon NA              
#> 5 Acipenser trans… NA               White Sturgeo… NA              
#> 6 Acipenser trans… NA               White Sturgeo… NA              
#> # … with 41 more variables: `Species Code` <chr>, `Element Code` <chr>, `Global
#> #   Status` <chr>, `Global Status Review Date` <chr>, `Prov Status` <chr>,
#> #   `Prov Status Review Date` <chr>, `Prov Status Change Date` <chr>,
#> #   COSEWIC <chr>, `COSEWIC Comments` <chr>, `BC List` <chr>, `Provincial
#> #   FRPA` <chr>, `Land Use Objectives` <lgl>, Ecosection <lgl>, `Prov Wildlife
#> #   Act` <lgl>, GOERT <lgl>, MBCA <lgl>, SARA <chr>, `SARA Comments` <chr>,
#> #   `General Status Canada` <chr>, CITES <chr>, `Name Category` <chr>, `Class
#> #   (English)` <chr>, `Species Level` <chr>, Kingdom <chr>, Phylum <chr>,
#> #   Class <chr>, Order <chr>, Family <chr>, `Forest Dist` <chr>, `MOE
#> #   Region` <chr>, `Regional Dist` <chr>, Municipality <chr>, BGC <chr>,
#> #   `Habitat Subtype` <chr>, Origin <chr>, Presence <chr>, `Breeding
#> #   Bird` <lgl>, `Ecosystem Group` <lgl>, Endemic <chr>, `CDC Maps` <chr>,
#> #   `Mapping Status` <chr>